requirements.txt
pandas
numpy
scikit-learn
scipy
fastapi
pyyaml
uvicorn
